Refactor and address code review feedback
