+++ helm-chart/dicom-adapter/files/appsettings.json
@@ -23,7 +23,7 @@
     "services": {
       "platform": {
         "uploadMetadata": false,
-        "metadata": [
+        "metadataDicomSource": [
           "0008,0020",
           "0008,0060",
           "0008,1030",
