+++ src/Configuration/ServicesConfiguration.cs
@@ -1,6 +1,6 @@
 ﻿/*
  * Apache License, Version 2.0
- * Copyright 2019-2020 NVIDIA Corporation
+ * Copyright 2019-2021 NVIDIA Corporation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,6 +16,7 @@
  */
 
 using System.Collections.Generic;
+using System.Runtime.Serialization;
 using Newtonsoft.Json;
 
 namespace Nvidia.Clara.DicomAdapter.Configuration
@@ -38,35 +39,5 @@ namespace Nvidia.Clara.DicomAdapter.Configuration
         /// <value></value>
         [JsonProperty(PropertyName = "resultsServiceEndpoint")]
         public string ResultsServiceEndpoint { get; set; }
-    }
-
-    public class PlatformConfiguration
-    {
-        /// <summary>
-        /// Gets or sets the URI of the Platform API.
-        /// </summary>
-        [JsonProperty(PropertyName = "endpoint")]
-        public string Endpoint { get; set; }
-
-        /// <summary>
-        /// Gets or sets maximum number of concurrent uploads to the Paylodas Service.
-        /// </summary>
-        [JsonProperty(PropertyName = "parallelUploads")]
-        public int ParallelUploads { get; set; } = 4;
-
-        /// <summary>
-        /// Gets or sets whether or not to upload metadata with the associated job.
-        /// If enabled, DICOM Adapter tries to extract all string and numeric fields specified in the
-        /// <c>MetadataDicomSource</c> field from one of the received DICOM instances.
-        /// </summary>
-        [JsonProperty(PropertyName = "uploadMetadata")]
-        public bool UploadMetadata { get; set; } = false;
-
-        /// <summary>
-        /// Gets or set a list of DICOM tags to be extracted and attached to the job triggered with the Clara Jobs Service.
-        /// </summary>
-        /// <value></value>
-        [JsonProperty(PropertyName = "metadataDicomSource")]
-        public List<string> MetadataDicomSource { get; set; } = new List<string>();
     }
 }
