+++ src/Server/Program.cs
@@ -140,8 +140,9 @@ namespace Nvidia.Clara.DicomAdapter
                     services.AddTransient<IResultsService, ResultsApi>();
                     services.AddTransient<IJobRepository, ClaraJobRepository>();
                     services.AddTransient<IInferenceRequestRepository, InferenceRequestRepository>();
-                    services.AddTransient(typeof(IDicomAdapterRepository<>), typeof(DicomAdapterRepository<>));
 
+                    services.AddScoped(typeof(IDicomAdapterRepository<>), typeof(DicomAdapterRepository<>));
+                    
                     services.AddSingleton<IStorageInfoProvider, StorageInfoProvider>();
                     services.AddSingleton<SpaceReclaimerService>();
                     services.AddSingleton<JobSubmissionService>();
