+++ src/Server/appsettings.Development.json
@@ -17,9 +17,9 @@
     },
     "services": {
       "platform": {
-        "endpoint": "192.168.20.209:7000",
+        "endpoint": "10.104.37.94:50051",
         "uploadMetadata": true,
-        "metadata": [
+        "metadataDicomSource": [
           "0008,0020",
           "0008,0060",
           "0008,1030",
